tests/test_web_interface_mode.py
import pytest
import sys
from unittest.mock import patch, MagicMock
import os

# Add src to path for imports
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))

def test_web_interface_mode():
    """Test that the web interface mode can be imported and started"""
    # This is a basic import test to ensure the web interface mode works
    try:
        from src.web import start_web_interface
        # Verify that the function exists
        assert callable(start_web_interface)
    except ImportError as e:
        pytest.fail(f"Failed to import web interface: {e}")
    except Exception as e:
        # This might fail due to missing dependencies in test environment
        # That's okay for this basic test
        pass

if __name__ == "__main__":
    test_web_interface_mode()
    print("Web interface mode test passed!")
